feat: improvement on generic errors management

+++ src/server/controllers/generic-controller/generic-controller.ts
@@ -1,5 +1,5 @@
 import { NextApiRouter } from "@/server/core/NextApiRouter";
-import { NotFoundError } from "@/server/helpers/api-error";
+import { BadRequestError, InternalError, NotFoundError } from "@/server/helpers/api-error";
 import { GENERIC_MESSAGES, TGenericMessage, TGenericMessageParams } from "@/server/references/generic-error";
 import { IGenericRepository } from "@/server/repositories/interfaces/generic-repository-intf";
 import IGenericController from "./generic-controller.intf";
@@ -29,6 +29,7 @@ export default class TGenericController<T> implements IGenericController {
     show() {
         this.$router.get(async (req, res) => {
             const { id } = req.query;
+            if (!id) throw new BadRequestError(this.$messages, { cases: 'required_field', contexts: 'repository', field: 'identificador' });
             const data = await this.$repository.findOne(id as string);
             if (!data) throw new NotFoundError(this.$messages, { cases: 'not_found', contexts: 'repository' });
             res.send(data)
@@ -49,18 +50,21 @@ export default class TGenericController<T> implements IGenericController {
         this.$router.put(async (req, res) => {
             const data = req.body;
             const { id } = req.query;
-            const updated = await this.$repository.updateOne({ id, ...data })
+            if (!id) throw new BadRequestError(this.$messages, { cases: 'required_field', contexts: 'repository', field: 'identificador' });
+            const updated = await this.$repository.updateOne({ id, ...data });
+            if (!updated) throw new InternalError(this.$messages, { cases: 'cannot_delete', contexts: 'repository', field: 'atualização' });
             res.send(updated);
         });
     }
     remove() {
         this.$router.delete(async (req, res) => {
             const { id } = req.query;
+            if (!id) throw new BadRequestError(this.$messages, { cases: 'required_field', contexts: 'repository', field: 'identificador' });
             const succeded = await this.$repository.deleteOne(id as string);
             if (succeded) {
                 res.status(200).end();
             } else {
-                res.status(400).end();
+                if (!id) throw new BadRequestError(this.$messages, { cases: 'cannot_delete', contexts: 'repository' });
             }
         })
     }
